requirements.txt
django
django-rest-framework
django-money
celery
redis